fix : main layout 수정 및 깨지는 css 수정
