--- src/pages/main/Main.module.css
@@ -1,18 +1,0 @@
-.sliderBig,
-.sliderMedium,
-.sliderSmall,
-.travelList,
-.hashtagList {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 50px 0;
-}
-.mainFirst {
-  margin: 0;
-  padding-top: 80px;
-}
-.sliderMedium,
-.hashtagList {
-  background: #f5f5f5;
-}
