+++ src/pages/main/components/bannerslide/BannerSlideMedium.module.css
@@ -1,9 +1,7 @@
 .container {
   position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
   width: 1200px;
+  margin: 0 auto;
 }
 
 .slideActive {
