+++ src/pages/main/components/list/HashtagList.module.css
@@ -1,5 +1,8 @@
 .hastagList {
+  display: flex;
+  justify-content: center;
   padding: 80px 0;
+  background: #f5f5f5;
 }
 .hashtagTopic {
   width: 100%;
